backend/requirements.txt
fastapi
uvicorn
pydantic_settings
pydantic[email]
passlib
python-jose
cryptography
bcrypt==4.0.1
python-multipart
motor
orjson
mutagen